+++ _content/recipes-old/beans-and-toast/beans-and-toast.mdx
