+++ _content/recipes-old/snickerdoodles/assets/cookies.png
