+++ _content/recipes/bread/focaccia/focaccia.mdx
@@ -1,6 +1,6 @@
 ---
 # General Information
-category: "Supper"
+category: "Bread"
 number: 1
 type: "Recipe"
 title: "Scale-Only No-Knead Focaccia"
